Move pin mapping to a separate file. SRP *and* we can apply a narrow SonarQube duplicate code suppression
